mckinsey_consultant_2/requirements.txt
crewai>=0.35.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
plotly>=5.15.0
streamlit>=1.28.0
python-pptx>=0.6.21
scikit-learn>=1.3.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.8.0
kaleido>=0.2.0
Pillow>=10.0.0
fpdf2>=2.7.7
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.10
